added custom hooks
