+++ src/utils/useRestaurantMenu.js
@@ -1,0 +1,19 @@
+import React, { useEffect } from "react";
+import { SWIGGY_MENU_URL } from "../utils/constant";
+
+const useRestaurantMenu = (resId) => {
+  const [resInfo, setResInfo] = useState(null);
+  useEffect(() => {
+    fetchMenu();
+  }, []);
+
+  const fetchMenu = async () => {
+    const data = await fetch(SWIGGY_MENU_URL + resId);
+    const json = await data.json();
+    // console.log(json);
+    setResInfo(json);
+  };
+  return resInfo;
+};
+
+export default useRestaurantMenu;
